Add the ability to further customize TLS connections
Previously, the only control over the TLS connection was
the ability to add new certificates with `Options::add_root_certificte`.
There are other config the user may be interested in setting. A new
method `Options::tls_config` was added to allow the user precise
control over the client's TLS configuration. `Options::native_certs`
was also added to allow a way to opt out of including the platforms
native certificate store.

Signed-off-by: David McNeil <mcneil.david2@gmail.com>