fix: add read and write timeouts to the tcp stream
The tcp stream to the nats-server can get stuck if the server is not
shutdown gracefully. Even the connection lost callback does not get
triggered because the stream is somehow stuck.

Adding the timeouts seems to resolve this issue, at least on my test
environment.

By default the timeouts are not set so the existing behavior is kept
but they can be set to a Duration value or None which is then set on
the tcp stream.