Bump nuid dependency to `0.3.1`
This is necessary as `0.3.0` was depending on `rand = "0"` thus it is possible for it to switch back to `0.7` or older, which just happened with `nuid` coming in transitively from our dependency on nats. The only change between `nuid 0.3.0` and `0.3.1` is the refinement of the `rand` version constraint. So I bet this is safe.