Fix jetstream PagedIterator infinite loop
On current nats server version (2.8) when a paged request is made to
jetstream (e.g. to list the stream names), the page will never be empty,
making the last page being repeated over and over and hanging if calling
collect() on it.
This fix implements the same fix as jsm.go(5a418c) manager.go +236 by
stopping when detecting it's the last page.