ref(kv): Changes `keys` method to use `Stream`
This is an optional commit that I figured I'd do while I was here.
The original documentation for this method suggested that it should
be returning a `Stream` rather than an iter. So I implemented a
stream version of it if desired. This commit can be popped off (or
moved to a separate PR) depending on maintainer preference

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>