Implement NatsError generic type
This type represents any error to be raised by the crate.
It is generic over the enum representing the kind of the error.

Unlike the previous code where binding between the kind and the error
used to be provided by the macro `error_impls`, this definition
makes the corresponding error bound to the kind more idiomatically
through the generic type parameter.

The previous syntax:

   #[derive(Clone, Debug, PartialEq)]
   pub enum FooErrorKind {
      Bar,
      Baz,
   }

   #[derive(Debug)]
   pub struct FooError {
      kind: FooErrorKind,
      source: Box<dyn Error + Send + Sync + 'static>,
   }

   error_impls!("FooError", "FooErrorKind");

The new definition:

   #[derive(Clone, Debug, PartialEq)]
   pub enum FooErrorKind {
      Bar,
      Baz,
   }

   pub type FooError = NatsError<FooErrorKind>;

In addition to being more idiomatic, the new syntax lets IDE-s
to understand the code better (most of them cannot analyze
the code inside the macro).