Implement Display for generic errors
From now and on it is enough to implement the Display
trait for the kind of the error. Then the Display trait
will be automatically implemented for the corresponding error.

In case the error contains the source, its formatting
will include both the kind and the source, otherwise
only the kind will be used for the display.

The method `formatted_source` is dropped.