Stop requiring error kinds to impl Display
All kinds of `error::Error` now implements the `Display` trait,
and the `Display` for the `Error` itself is inherited.

The binding between the kind and `Display` is not necessary though.

In this commit I dropped the restriction of error kinds by `Display`,
and only implemented the trait for errors with displayable kinds.
As a result, the developer got 2 options:

- either to implement the `Display` for the kind (preferred),
- or implement it for the error only (but not for the kind) (edge case).