Turn off incremental compilation
The incremental build is a way to fasten future compilations
for the cost of extra dependency tracking and bigger /target.

This <default> approach is not necessary in the CI
because every compilation is done from scratch.

As a result of turning it off the size of the target
reduces from 660MB to 543MB.