Allow async and sync nats to compile on MIPS and PowerPC
Now that nats.rs depends on a version of rustls that includes the latest
version of ring, it can now be compiled to run on additional types of
CPU architectures.

Specifically this converts all uses of `AtomicU64` to use the type
provided by the `portable-atomic` crate, which uses the native version
of `AtomicU64` on all architectures except ones where that type isn't
supported.

I went ahead and updated sync nats with the latest version of rustls
which inflated the size and scope of this PR beyond just swapping out
the AtomicU64 implementation.