Don't expose BoxFuture in public API
`ObjectStore::get` calls itself recursively, so the returned future
needs to be boxed. However, this is an implementation detail, there's no
reason to expose it in the public API. Instead, we can move the
recursive call to a separate `get_impl` function that handles the
boxing, and `get` can just be a standard async function.