fix: ensure the ping interval waker is registered
When an interval is ready it does not need to register its internal
waker, leading to the future not being polled anymore if no data is
received from the connection.

The fix is pretty simple, and it should never cause any problem unless
the interval duration is very close to zero, which will probably cause
strange issues anyway.