Remove dependency for `async-nats` from `nats-server`.
Without this change during Cargo sync step, the following error is produced:
```
error: failed to select a version for `async-nats`.
    ... required by package `nats-server v0.1.0 (\github\nats.rs\nats-server)`
    ... which satisfies path dependency `nats-server` (locked to 0.1.0) of package `nats v0.25.0 (\github\nats.rs\nats)`
versions that meet the requirements `^0.38` (locked to 0.38.0) are: 0.38.0
package `nats-server` depends on `async-nats` with feature `server_2_11` but `async-nats` does not have that feature.
 package `async-nats` does have feature `server_2_10`
failed to select a version for `async-nats` which could resolve this conflict
```